com.javarush.test.level10.lesson11.home02 changed.

+++ com/javarush/test/level10/lesson11/home02/Solution.java
@@ -1,19 +1,38 @@
 package com.javarush.test.level10.lesson11.home02;
 
 /**
- * User: General
- * Date: 12/21/12
- * Time: 11:59 PM
+ * JavaRush.ru
+ * Level 10, Lesson 11, Home 02
+ * <p/>
+ * Добавить одну операцию по преобразованию типа, чтобы получался ответ: d = 5.5
+ * <p/>
+ * Date: 19.04.13
+ * @author Sergey Kandalintsev
  */
 public class Solution
 {
-    public static void main(String[] args)
+    public static void main( String[] args )
     {
         int a = 5;
         int b = 4;
         int c = 3;
         int e = 2;
-        double d = a + b/c/e;
-        System.out.println(d);
+        double d = a + b / c / ( double ) e;
+
+        System.out.println( d );
     }
 }
+
+//Исходный вариант:
+//public class Solution
+//{
+//    public static void main( String[] args )
+//    {
+//        int a = 5;
+//        int b = 4;
+//        int c = 3;
+//        int e = 2;
+//        double d = a + b / c / e;
+//        System.out.println( d );
+//    }
+//}
